broker/inference/aws/requirements.txt
tensorflow
numpy
pillow
boto3
